Update myStyle.scss
Added and integrated a Controlled Panel Accordion using React-Bootstrap.js. Because react-bootstrap.js is only working with Bootstrap.css v.3, and the last design was based on Bootstrap.css v.4, I had to revisit and optimize the style of the modal elements and all the buttons. I integrated the "Delete" and "Edit" buttons also in the Panel Accordion.