Update myScript.js
Added and integrated a Controlled Panel Accordion using React-Bootstrap.js. Now every recipe entry works well and has it's own separate panel, where a user can see the name of the recipe and the ingredients. As a little extra I added a counter on the footer of the panel that keeps count on the number of the added recipes. I integrated the "Delete" and "Edit" buttons also, just need to add the handle methods for both and the project is completed.