Update myScript.js
Implemented a method that handles the delete button. On the onClick event it retrieves the name of the recipe and sets in state then it's passed as a prop to the final method/function that handles the removal of the entry from the recipe box. Next I need to implement a filter method that rejects duplicates on a new recipe add and then create the method for editing the recipe.