Update myScript.js
Implemented the function that blocks the duplicate entries on new recipe add. Also added the handleEdit() method that at this moment when the "Edit" button is clicked, loads the Modal dialog window and injects the Name and Ingredients of that particular recipe, in their respective input and textarea fields. Now on "Close" button click the method behind it now clears the state that controls the the input fields. Need to modify the Header of the Modal dialog window on edit call, to show that you are editing an existing recipe. Also the "add recipe" button needs to become "Save".