Update myScript.js
The App is almost complete. As an extra, I've written a function that prompts the user to add a recipe name ( using a color highlight ) in the case when the "+ Recipe" button is clicked and the input field for the recipe name is empty. This avoids adding an "untitled recipe" in the box. Implemented a method that sets and gets the recipes array to and from localStorage and sets the state on componentDidMount(). Need to add a storageUpdater() function that handles the updates made to the recipes arr in the localStorage and all is complete.