Update myStyle.scss
Made a few changes to the button styles. Set a more clear font-family for the entire body of the app to add a more uniform experience. For a fresher look I replicated the feel of the buttons from Bootstrap v.4, although I'm using Bootstrap v.3 needed by the React-Bootstrap components. The next commit will be for the final stage of the project.